+++ booklog/src/components/Body/Books/Books.tsx
@@ -1,7 +1,7 @@
 import { useContext } from "react";
 import { BooksContext } from "../Body";
 import { Book } from "./Book/Book";
-import type { BooksResult, BookItem } from "../../../types";
+import type { BooksResult, BookItem, MyBooksContextType } from "../../../types";
 import styles from "./Books.module.css";
 
 type Props = {
@@ -10,19 +10,7 @@ type Props = {
 };
 
 export const Books = ({ bookItems, total }: Props): JSX.Element => {
-  const { myBooks, setMyBooks } = useContext(BooksContext);
-
-  function AddMyBooks(id: string) {
-    const selectedBook = bookItems.find((item) => item.id === id);
-    if (selectedBook) {
-      const checkAvailable = myBooks.find((item) => item.id === id);
-      if (!checkAvailable) {
-        setMyBooks([...myBooks, selectedBook]);
-      } else {
-        alert("その本はすでにマイブックに存在します。");
-      }
-    }
-  }
+  const { AddMyBooks } = useContext(BooksContext) as MyBooksContextType;
 
   return (
     <>
