+++ booklog/src/components/Body/SideBar/SideBar.module.css
@@ -10,10 +10,14 @@
   font-size: 16px;
 }
 
+.conButton {
+  display: flex;
+  margin-top: 10px;
+}
+
 .boxButton {
   display: inline-block;
   padding: 3px 10px;
-  margin-top: 10px;
   opacity: 1;
   border-radius: 5px;
   font-size: 14px;
